refactor: change action chain build algorithm to iterate on actions instead of collections
Some Actions are needed to specify a place where they should be in action chain. For example CreateEmbedded action should be in very top of chain since other actions may refer to embedded document which it creates. Another example is that RenameCollection must be checked before CreateCollection/DropCollection because collection renaming looks like dropping old and creating new one. And so forth. So this commit changes iteration algorithm to iterate over collections with every actions instead of iterate over actions with every collection.