refactor: extract all variables related to migartions run to a "run context"
The problem is that we could be confused when we use Action variables which are filled only when migrations has begun. We could mix up them with variables filled on Action constructor. So it's better to extract them to separate place and denote somehow that they are valid only when migration is running. So there is "run context"